requirements.txt
amqp==5.2.0
asgiref==3.8.1
billiard==4.2.0
black==24.4.2
celery==5.4.0
certifi==2024.7.4
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
Django==5.0.7
django-restframework==0.0.1
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
elastic-transport==8.13.1
elasticsearch==8.14.0
elasticsearch-dsl==8.14.0
flake8==7.1.0
flower==2.0.1
gunicorn==22.0.0
humanize==4.10.0
isort==5.13.2
kombu==5.3.7
mccabe==0.7.0
mypy-extensions==1.0.0
packaging==24.1
pathspec==0.12.1
pillow==10.4.0
platformdirs==4.2.2
prometheus_client==0.20.0
prompt_toolkit==3.0.47
psycopg2==2.9.9
pycodestyle==2.12.0
pyflakes==3.2.0
PyJWT==2.8.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
six==1.16.0
sqlparse==0.5.0
tornado==6.4.1
tzdata==2024.1
urllib3==2.2.2
vine==5.1.0
wcwidth==0.2.13
